NONE: refactor mainpage
